Update README.md
Main Goal: Implement a queue both using an array and as a linked list!
Problem to solve:
A doctor is providing 10 seats in the waiting room for clients to be
served in FIFO! For this you will have to use a queue called 
clients_queue. 
We will prioritize clients with an emergency case and for that we will 
use a queue called emergency_client_queue. 
Important points to consider:
- For this project you will have to use two structures one named 
client_info (consists of 2 elements: name, and a ticket number) 
and a hierarchical structure called urgent_client_info (consists of 
two elements: client_info element and a next pointer)
- You will have to use 2 global variables called ticket_number_n 
and ticket_number_u that you should initialize to zero
Let’s use a menu that consists of the following options:
1. Receive a New Client
2. Serve Client
3. Check a Queue Content
4. Change the Status of a Client
5. Service is OVER
- For the first menu choice, it is meant for adding a client to a queue: 
o You need to ask the user for the status: normal case or 
emergency case 
o You need a function that will be used to fill the client data
o You need to call the appropriate enqueue function 
- For the second option, it is meant to serve a client. Please note 
priority is given to clients with an emergency. Unless the queue of 
emergencies is empty, a client with no emergency will be served.
- The third option is meant to display the content of a queue
o Need to ask the user for which queue to check
- For the fourth case: in some cases, a client can show up for a normal 
visit which end up after to an emergency. Let’s define a function that 
will remove a client X from the first queue and enqueue him/her to 
the emergency queue. 
o Note that node should be created for this client
o If the client is not found in the queue, display an error in the 
main function
- When the user decides to quit, make sure that the emergency 
queue is empty, otherwise don’t end the program. Clients that are
waiting in the non-emergency queue will be asked to visit the 
doctor next time. Using a function, your output should be: 
Dear client1(print the name), client2(print the name),……, Sorry to 
inform you that the doctor can receive another day.